fixed delete macro usage fixed signed/unsigned usage documentation changelog more fixes on dist-test.sh
git-svn-id: https://clucene.svn.sourceforge.net/svnroot/clucene/branches/lucene2_3_2@2868 20ef185c-fe11-0410-a618-ba9304b01011

+++ src/shared/CLucene/config/repl_tprintf.cpp
@@ -140,7 +140,7 @@ void lucene_wprintf(const wchar_t * format, ...){
 int lucene_vsnwprintf(wchar_t * strbuf, size_t count, const wchar_t * format, va_list& ap){
 	StringBuffer buffer;
     lucene_vfnwprintf(&buffer,count,format,ap);
-	int ret = cl_min((int32_t)count,buffer.length()+1);
+	int ret = cl_min((size_t)count,buffer.length()+1);
 	wcsncpy(strbuf,buffer.getBuffer(),ret);
     return ret;
 }
